requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
dataclasses-json==0.6.7
diskcache==5.6.3
distro==1.9.0
fastapi==0.128.0
filetype==1.2.0
fpdf==1.7.2
frozenlist==1.8.0
google-auth==2.48.0
google-genai==1.63.0
greenlet==3.3.1
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==1.2.8
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.14
langchain-google-genai==4.2.0
langchain-groq==1.1.2
langchain-ollama==1.0.1
langchain-openai==1.1.10
langchain-text-splitters==1.1.0
langgraph==1.0.7
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4
langsmith==0.6.9
llama_cpp_python==0.3.16
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.3
ollama==0.6.1
openai==2.21.0
opencv-python==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
propcache==0.4.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
starlette==0.50.0
tenacity==9.1.3
tiktoken==0.12.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
